fix(render): root-cause render pipeline — 5-layer failure chain
Hook install was GREEN but nothing drew. Root causes found by code
audit (log showed only the install chain, drawing errors were invisible
because CoreLogger wrote to stdout only):

1. MappingContext.invokeMethod: bindTo(null) on zero-arg static
   MethodHandle (Minecraft.getMinecraft) throws 'no leading reference
   parameter' -> mc was null on EVERY call -> ForgeBootstrap.render()
   silently returned at first check. Fix: invoke directly when obj==null.
2. mappings/forge/v1_8_9.json: forge:mc_thePlayer used 1.7.x SRG name
   field_70475_c; 1.8.9 runtime name is field_71439_g (log-verified).
   player was always null -> render() returned at second check.
3. mappings/forge/v1_8_9.json: fontRenderer_drawStringWithShadow used
   func_78261_a (1.7.x int-signature); 1.8.9 uses func_175063_a
   (float-signature, log-verified) -> text never drew.
4. ForgeGL11Bridge: glColor4f/glVertex2f/glDepthMask reflection used
   boxed Float::class.java/Boolean::class.java -> NoSuchMethodException
   -> all rect drawing silently failed. Fix: javaPrimitiveType.
5. ForgeFontRendererBridge: pass x/y as float to match func_175063_a
   signature.
Plus: CoreLogger now mirrors to %TEMP%\switchlite-agent.log so render
diagnostics are visible in the log dump instead of lost stdout.

+++ SwitchLite/adapter/forge/v1_8_9/src/main/kotlin/io/switchlite/adapter/forge/v1_8_9/ForgeFontRendererBridge.kt
@@ -42,7 +42,7 @@ class ForgeFontRendererBridge(private val fontRendererObj: Any) : FontRendererBr
                 }
                 return 0
             }
-            handle.bindTo(fontRendererObj).invokeWithArguments(text, x, y, color) as? Int ?: 0
+            handle.bindTo(fontRendererObj).invokeWithArguments(text, x.toFloat(), y.toFloat(), color) as? Int ?: 0
         } catch (e: Exception) {
             if (!drawStringErrorLogged) {
                 CoreLogger.error("[ForgeFontRendererBridge] drawStringWithShadow failed: ${e.javaClass.simpleName}: ${e.message}")
